Large changes to pool functionality
Pool functionality was changed to remove blocking when running jobs with
pools. Pools no longer return results directly. All pending batches
of jobs are instead first cached to a queue. The user then chooses
when to retrieve the results; all retrieved results are stored in a pool
results queue and all pending results to that point are returned to the
user.